user globar var for parsing json

+++ views/users.py
@@ -1,27 +1,24 @@
-import sqlite3
 from models.users import UserModel
 from flask_restful import Resource,reqparse
-from flask_jwt_extended import (
-    create_access_token,
-    create_refresh_token,
+from flask_jwt_extended import create_access_token,create_refresh_token
+from flask_jwt_extended import jwt_required, get_jwt
 
-)
+_user_parser = reqparse.RequestParser()
+_user_parser.add_argument('username',
+        type=str,
+        required=True,
+        help="this field not can be blank"
+        )
+_user_parser.add_argument('password',
+        type=str,
+        required=True,
+        help="this field not can be blank"
+        )
+
 
 class UserRegister(Resource):
-    def __init__(self):
-        self.parser = reqparse.RequestParser()
-        self.parser.add_argument('username',
-        type=str,
-        required=True,
-        help="this field not can be blank"
-        )
-        self.parser.add_argument('password',
-        type=str,
-        required=True,
-        help="this field not can be blank"
-        )
     def post(self):
-        data = self.parser.parse_args()
+        data = _user_parser.parse_args()
         if UserModel.find_by_username(data['username']):
             return {"MSG":"We have this user naem"}
         user = UserModel(**data)
@@ -29,14 +26,17 @@ class UserRegister(Resource):
         return {"msg":"user was create"}
 
 class User(Resource):
-    @classmethod
+
     def get(cls,user_id):
         user = UserModel.find_by_id(user_id)
         if not user:
             return {"msg": "user nOt found "},404
         return user.json()
-    @classmethod
+
     def delete(cls,user_id):
+        claims = get_jwt()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required.'}, 401
         if UserModel.find_by_id(user_id):
             UserModel.find_by_id(user_id).delete_from_db()
             return{"msg":"user was deleted"}
@@ -44,21 +44,10 @@ class User(Resource):
             return{"msg":"user not found"}
         
 class UserLogin(Resource):
-    def __init__(self):
-        self.parser = reqparse.RequestParser()
-        self.parser.add_argument('username',
-        type=str,
-        required=True,
-        help="this field not can be blank"
-        )
-        self.parser.add_argument('password',
-        type=str,
-        required=True,
-        help="this field not can be blank"
-        )
 
-    def post(self):
-        data = self.parser.parse_args()
+
+    def post(cls):
+        data = _user_parser.parse_args()
         user = UserModel.find_by_username(data['username'])
         if user and data['password']:
             access_token = create_access_token(identity= user.id,fresh=True)
@@ -66,5 +55,5 @@ class UserLogin(Resource):
             return{
                 "access_token":access_token,
                 "refresh_token":refresh_token,
-            }
+            },200
         return {"msg": "your username or password was not correct"}
